Made the analysis strategy and the loop-upper bound for indefinable loops variable.

+++ src/main/include/LLVM-Handler/LLVMHandler.cpp
@@ -2,7 +2,7 @@
 #include "LLVMHandler.h"
 #include "InstructionCategory.h"
 
-LLVMHandler::LLVMHandler( Json::Value energy, long valueIfIntederminate) {
+LLVMHandler::LLVMHandler( Json::Value energy, int valueIfIntederminate) {
     this->energyValues = energy;
     this->valueIfIndeterminate = valueIfIntederminate;
 }
